# contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Internal Audience & Campaign Control API
  version: 0.8.28
  description: >
    Internal-only API for organisation records, offline consent reviews, secure audience imports,
    governed cohort definitions and controlled campaign lifecycle management.
servers:
  - url: /api/v1
security:
  - internalSession: []
paths:
...
  /consent-events:
    get:
      operationId: listConsentEvents
      summary: listConsentEvents
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Request completed
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /consent-grants:
    post:
      operationId: createConsentGrant
      summary: createConsentGrant
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /consent-grants/{id}/withdraw:
    post:
      operationId: withdrawConsentGrant
      summary: withdrawConsentGrant
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
...
  /suppressions:
    post:
      operationId: createSuppression
      summary: createSuppression
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /suppressions/{id}/revoke:
    post:
      operationId: revokeSuppression
      summary: revokeSuppression
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
...
  /contacts/{id}/lifecycle:
    get:
      operationId: getContactLifecycle
      summary: Get the protected lifecycle status for one contact
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Contact lifecycle status }
        '404': { description: Contact not found }
    post:
      operationId: transitionContactLifecycle
      summary: Apply an MFA-protected and optimistically locked contact lifecycle transition
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, processingRestricted, expectedVersion, reason]
              properties:
                status: { type: string, enum: [ACTIVE, INACTIVE, SUPPRESSED, INVALID, RECYCLED, DECEASED] }
                processingRestricted: { type: boolean }
                expectedVersion: { type: integer, minimum: 1 }
                reason: { type: string, minLength: 12 }
      responses:
        '200': { description: Lifecycle transition recorded }
        '403': { description: Permission or MFA step-up required }
        '404': { description: Contact not found }
        '409': { description: Version conflict or immutable privacy state }
        '422': { description: Invalid lifecycle transition }
  /contacts/{id}/lifecycle/events:
    get:
      operationId: listContactLifecycleEvents
      summary: List immutable contact lifecycle events
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Contact lifecycle event list
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
        '404': { description: Contact not found }
...
components:
  securitySchemes:
    internalSession:
      type: apiKey
      in: cookie
      name: campaign_session
...